pyproject.toml
[tool.poetry]
name = "relayer"
version = "0.1.0"
description = ""
authors = ["spourdei <spourdei@uwaterloo.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
